Convert to apiInitializer and move to api-initializers/
The old withPluginApi/initialize(container) pattern in the initializers/
directory is deprecated for themes. Converts to apiInitializer in the
api-initializers/ directory, the modern equivalent.
